Update README.md
The Sesotho language is one of the previously disadvantaged languages of Africa spoken in Lesotho and South Africa. This project is based on compiling a corpus for Sesotho and developing a frequency list that will assist lexicographers of learner dictionaries of which words to lemmatize based on the results of the words that are most likely to be looked for by users. This approach takes a user perspective in mind.